requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atlassian-python-api==4.0.4
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
Deprecated==1.2.18
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
idna==3.10
jiter==0.10.0
jmespath==1.0.1
lmstudio==1.3.1
msgspec==0.19.0
oauthlib==3.2.2
openai==1.86.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.4
requests-oauthlib==2.0.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wrapt==1.17.2
wsproto==1.2.0
